+++ src/FixedWidthParser/Readers/FixedWidthByteReader.cs
@@ -24,7 +24,7 @@ namespace FixedWidthParser.Readers
         private readonly StringPool? _stringPool;
         private readonly int _bufferSize;
 
-        public FixedWidthByteReader(IFormatProvider? formatProvider = null, StringPool? stringPool = null, int bufferSize = 4096)
+        public FixedWidthByteReader(IFormatProvider? formatProvider = null, StringPool? stringPool = null, int bufferSize = 65536)
         {
             ArgumentOutOfRangeException.ThrowIfLessThan(bufferSize, 1);
             _formatProvider = formatProvider;